Add shim for ubnthal
This change adds a shim implementation of the proprietary ubnthal kernel
module shipped with the firmware on the UniFi Dream Machine. This
open-source shim facilitates applications depending on the data exposed
by ubnthal, without requiring this module to run, and makes it easier to
use custom kernels with UniFi firmware.